# Conflicts:
#	pom.xml

+++ src/test/java/org/mariadb/jdbc/ResultSetUnsupportedMethodsTest.java
@@ -3,7 +3,7 @@
  * MariaDB Client for Java
  *
  * Copyright (c) 2012-2014 Monty Program Ab.
- * Copyright (c) 2015-2017 MariaDB Ab.
+ * Copyright (c) 2015-2019 MariaDB Ab.
  *
  * This library is free software; you can redistribute it and/or modify it under
  * the terms of the GNU Lesser General Public License as published by the Free
@@ -57,7 +57,6 @@ import java.sql.SQLException;
 import java.sql.SQLFeatureNotSupportedException;
 import org.junit.Before;
 import org.junit.Test;
-
 
 public class ResultSetUnsupportedMethodsTest extends BaseTest {
 
@@ -240,12 +239,12 @@ public class ResultSetUnsupportedMethodsTest extends BaseTest {
 
   @Test(expected = SQLFeatureNotSupportedException.class)
   public void testupdateLong10() throws SQLException {
-    rs.updateLong(1, (long) 1);
+    rs.updateLong(1, 1);
   }
 
   @Test(expected = SQLFeatureNotSupportedException.class)
   public void testupdateLong11() throws SQLException {
-    rs.updateLong("a", (long) 1);
+    rs.updateLong("a", 1);
   }
 
   @Test(expected = SQLFeatureNotSupportedException.class)
@@ -260,12 +259,12 @@ public class ResultSetUnsupportedMethodsTest extends BaseTest {
 
   @Test(expected = SQLFeatureNotSupportedException.class)
   public void testupdateDouble14() throws SQLException {
-    rs.updateDouble(1, (double) 1);
+    rs.updateDouble(1, 1);
   }
 
   @Test(expected = SQLFeatureNotSupportedException.class)
   public void testupdateDouble15() throws SQLException {
-    rs.updateDouble("a", (double) 1);
+    rs.updateDouble("a", 1);
   }
 
   @Test(expected = SQLFeatureNotSupportedException.class)
@@ -597,6 +596,4 @@ public class ResultSetUnsupportedMethodsTest extends BaseTest {
   public void testupdateNCharacterStream81() throws SQLException {
     rs.updateNCharacterStream("a", null);
   }
-
-
 }
